+++ examples/randomNoise.js
@@ -2,13 +2,13 @@ import { Canvas } from "../index.js";
 const width = 300;
 const height = 300;
 const canvas = new Canvas("Random noise", width, height);
-while (true) {
-    let buf = randomData();
+let buf = new Uint8Array(width * height);
+canvas.loop(() => {
+    randomData();
     canvas.loadRawData(buf, 8);
-}
+})
 
 function randomData() {
-    let buff = new Uint8Array(width * height);
-    for (let i = 0; i < buff.length; i++) buff[i] = Math.floor(Math.random() * 255);
-    return buff;
+    for (let i = 0; i < buf.length; i++) buf[i] = Math.floor(Math.random() * 255);
+    return buf;
 }